requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
requests>=2.32.0
pytest>=8.0.0
pyyaml>=6.0.0
